+++ gradle.properties
@@ -7,14 +7,8 @@ org.gradle.daemon = false
 tweaked_lib_version = 1.0.0
 
 mod_version = 1.0.1
-maven_group = srki2k
+maven_group = io.github.srdjanv
 archives_base_name = tweakedexcavation
-
-
-# Optionial Jar compiling
-build_deobfJar = false
-build_apiJar = false
-build_sourceJar = true
 
 # If any properties changes below this line, run `gradlew setupDecompWorkspace` and refresh gradle again to ensure everything is working correctly.
 
@@ -22,8 +16,10 @@ build_sourceJar = true
 use_mixins = true
 use_coremod = true
 use_assetmover = false
-has_access_transformer = false
+
+# Access Transformer files should be in the root of `resources` folder and with the filename formatted as: `{archives_base_name}_at.cfg`
+use_access_transformer = false
 
 # Coremod Arguments
 include_mod = true
-coremod_plugin_class_name = srki2k.tweakedexcavation.core.TweakedExcavationPlugin
+coremod_plugin_class_name = io.github.srdjanv.tweakedexcavation.core.TweakedExcavationPlugin
