+++ settings.gradle
@@ -1,17 +1,22 @@
 pluginManagement {
     repositories {
+        maven {
+            // RetroFuturaGradle
+            name = "GTNH Maven"
+            url = uri("http://jenkins.usrv.eu:8081/nexus/content/groups/public/")
+            allowInsecureProtocol = true
+            mavenContent {
+                includeGroup("com.gtnewhorizons")
+                includeGroup("com.gtnewhorizons.retrofuturagradle")
+            }
+        }
         gradlePluginPortal()
-        maven {
-            name 'FancyGradle'
-            url 'https://gitlab.com/api/v4/projects/26758973/packages/maven'
-        }
-        maven {
-            name 'ForgeGradle 5'
-            url 'https://maven.minecraftforge.net'
-        }
+        mavenCentral()
+        mavenLocal()
     }
 }
 
 plugins {
-    id 'com.alexvasilkov.git-dependencies' version '2.0.4'
+    // Automatic toolchain provisioning
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
 }
