+++ src/main/java/io/github/srdjanv/tweakedexcavation/api/ihelpers/IExcavatorAddons.java
@@ -1,0 +1,11 @@
+package io.github.srdjanv.tweakedexcavation.api.ihelpers;
+
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+public interface IExcavatorAddons {
+
+    void initEnergyStorage();
+
+    void onUpdate(CallbackInfo ci);
+
+}
