+++ src/main/java/io/github/srdjanv/tweakedexcavation/api/ihelpers/IMineralMix.java
@@ -1,0 +1,15 @@
+package io.github.srdjanv.tweakedexcavation.api.ihelpers;
+
+public interface IMineralMix {
+
+    int getPowerTier();
+
+    void setPowerTier(int powerTier);
+
+    int getYield();
+
+    void setYield(int yield);
+
+}
+
+
