+++ src/main/java/io/github/srdjanv/tweakedexcavation/api/ihelpers/IMineralWorldInfo.java
@@ -1,0 +1,9 @@
+package io.github.srdjanv.tweakedexcavation.api.ihelpers;
+
+import blusunrize.immersiveengineering.api.tool.ExcavatorHandler;
+
+public interface IMineralWorldInfo {
+    ExcavatorHandler.MineralMix getType();
+
+    int getDepletion();
+}
