+++ src/main/java/io/github/srdjanv/tweakedexcavation/common/DefaultMineralPower.java
@@ -1,0 +1,28 @@
+package io.github.srdjanv.tweakedexcavation.common;
+
+import io.github.srdjanv.tweakedlib.api.powertier.PowerTierHandler;
+
+public class DefaultMineralPower {
+
+    public static DefaultMineralPower instance;
+
+    public static DefaultMineralPower getInstance() {
+        if (instance == null) {
+            instance = new DefaultMineralPower();
+        }
+        return instance;
+    }
+    private DefaultMineralPower() {
+        powerTier = PowerTierHandler.registerPowerTier(
+                Configs.TPConfig.DefaultExcavatorPowerTiers.capacity,
+                Configs.TPConfig.DefaultExcavatorPowerTiers.rft
+        );
+    }
+
+    private final int powerTier;
+
+    public int getPowerTier() {
+        return powerTier;
+    }
+
+}
