--- src/main/java/srki2k/tweakedexcavation/api/ihelpers/IExcavatorAddons.java
@@ -1,11 +1,0 @@
-package srki2k.tweakedexcavation.api.ihelpers;
-
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-public interface IExcavatorAddons {
-
-    void initEnergyStorage();
-
-    void onUpdate(CallbackInfo ci);
-
-}
