--- src/main/java/srki2k/tweakedexcavation/api/ihelpers/IMineralMix.java
@@ -1,15 +1,0 @@
-package srki2k.tweakedexcavation.api.ihelpers;
-
-public interface IMineralMix {
-
-    int getPowerTier();
-
-    void setPowerTier(int powerTier);
-
-    int getYield();
-
-    void setYield(int yield);
-
-}
-
-
