--- src/main/java/srki2k/tweakedexcavation/api/ihelpers/IMineralWorldInfo.java
@@ -1,9 +1,0 @@
-package srki2k.tweakedexcavation.api.ihelpers;
-
-import blusunrize.immersiveengineering.api.tool.ExcavatorHandler;
-
-public interface IMineralWorldInfo {
-    ExcavatorHandler.MineralMix getType();
-
-    int getDepletion();
-}
