+++ src/main/resources/mixins.tweakedexcavation.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "srki2k.tweakedexcavation.mixin",
+  "package": "io.github.srdjanv.tweakedexcavation.mixin",
   "refmap": "mixins.tweakedexcavation.refmap.json",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
